- added login; - added jwt token;

+++ backend/controllers/feed.js
@@ -2,6 +2,7 @@ const fs = require('fs');
 const path = require('path');
 const { validationResult } = require('express-validator');
 const Post = require('../models/post');
+const User = require('../models/user');
 
 const ITEMS_PER_PAGE = 2;
 
@@ -20,6 +21,7 @@ exports.getPosts = (req, res, next) => {
         .limit(perPage);
     })
     .then(posts => {
+      // res.set('Cache-Control', 'no-store, no-cache, must-revalidate');
       res.status(200).json({
         message: 'Posts fetched successfully.',
         posts: posts,
@@ -39,6 +41,7 @@ exports.createPost = (req, res, next) => {
   if (!errors.isEmpty()) {
     const error = new Error('Validation failed, entered data is incorrect.');
     error.statusCode = 422;
+    error.data = errors.array();
     throw error;
   }
 
@@ -51,18 +54,29 @@ exports.createPost = (req, res, next) => {
   const imageUrl = req.file.path;
   const title = req.body.title;
   const content = req.body.content;
+  let creator;
+  // const user = await User.findById(creator);
   const post = new Post({
     title: title,
     content: content,
     imageUrl: imageUrl,
-    creator: { name: 'Art' }
+    creator: req.userId
   });
   
   post.save()
     .then(result => {
+      return User.findById(result.creator);
+    })
+    .then(user => {
+      creator = user;
+      user.posts.push(post);
+      return user.save();
+    })
+    .then(result => {
       res.status(201).json({
         message: 'Post created successfully',
-        post: result,
+        post: post,
+        creator: { _id: creator._id, name: creator.name }
       });
     })
     .catch(err => {
@@ -113,9 +127,10 @@ exports.updatePost = (req, res, next) => {
     throw error;
   }
 
-  if(!errors.isEmpty()) {
+  if (!errors.isEmpty()) {
     const error = new Error('Validation failed, entered data is incorrect.');
     error.statusCode = 422;
+    error.data = errors.array();
     throw error;
   }
 
@@ -126,6 +141,13 @@ exports.updatePost = (req, res, next) => {
         error.statusCode = 404;
         throw error;
       }
+
+      if(post.creator.toString() !== req.userId) {
+        const error = new Error('Not authorized!');
+        error.statusCode = 403;
+        throw error;
+      }
+
       if(post.imageUrl !== imageUrl) {
         clearImage(post.imageUrl);
       }
@@ -158,8 +180,22 @@ exports.deletePost = (req, res, next) => {
         error.statusCode = 404;
         throw error;
       }
+
+      if(post.creator.toString() !== req.userId) {
+        const error = new Error('Not authorized!');
+        error.statusCode = 403;
+        throw error;
+      }
+
       clearImage(post.imageUrl);
       return Post.findByIdAndDelete(postId);
+    })
+    .then(result => {
+      return User.findById(req.userId);
+    })
+    .then(user => {
+      user.posts.pull(postId);
+      return user.save();
     })
     .then(result => {
       res.status(200).json({
